Move traffic shaping plugin to traffic shaping module. Move APQ and dedup to traffic shaping.
