+++ apollo-router-core/src/layers/mod.rs
@@ -1,9 +1,7 @@
 //! Reusable layers
 //! Layers that are specific to one plugin should not be placed in this module.
-pub mod apq;
 pub mod async_checkpoint;
 pub mod cache;
-pub mod deduplication;
 pub mod ensure_query_presence;
 pub mod forbid_http_get_mutations;
 pub mod instrument;
