+++ apollo-router-core/src/plugins/mod.rs
@@ -6,4 +6,4 @@ mod forbid_mutations;
 mod headers;
 mod include_subgraph_errors;
 pub mod serde_utils;
-mod traffic_shaping;
+pub(crate) mod traffic_shaping;
